requirements.txt
asgiref==3.2.7
Django==3.0.6
django-cors-headers==3.4.0
django-filter==2.2.0
djangorestframework==3.11.0
importlib-metadata==1.6.0
Markdown==3.2.2
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
